# openapi.yaml
openapi: 3.0.3
info:
  title: Alauda API
  version: 1.0.0
  description: |
    🚀 **Alauda API - Multi-Service Integration Platform**
    
    All-in-one API for media downloads, social media content, music identification, payment processing and more. Built for developers who need reliable, fast and affordable integrations.
    
    ## 🎯 Features
    
    **Media Downloads:**
    - 🎵 Spotify - Search and download tracks
    - 📹 YouTube - Video and audio downloader
    - 🎬 TikTok - Download videos without watermark
    - 📸 Instagram - Download posts, reels, stories
    - 📘 Facebook - Video downloader
    
    **Music Services:**
    - 🎤 Shazam - Identify any song
    - 📝 Lyrics - Search song lyrics worldwide
    
    **Utilities:**
    - 🆔 CPF Validator - Brazilian tax ID validation
    - 🖼️ Background Remover - AI-powered image processing
    - 💬 WhatsApp Utils - Formatting and validation tools
    
    **Payment Processing:**
    - 💳 MercadoPago - Brazil and LATAM payments
    - 📱 M-Pesa - Mozambique mobile money (Vodacom)
    - 📱 E-Mola - Mozambique mobile money (Movitel)
    
    ## 🌍 Developed in Mozambique 🇲🇿 | Available Worldwide
    
    **Created by:** Zeus Lykraios 💎
    
    ## 🔑 Authentication
    
    All endpoints (except `/` and `/health`) require an API key sent in the `X-API-Key` header.
    
    Get your free API key at: https://alauda-api.topazioverse.com.br/api/auth/register
    
    ## 💰 Credits System
    
...
    
    ## 📊 Rate Limiting
    
...
    
  contact:
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/tiktok/download:
    post:
      tags:
        - Media Downloads
      summary: Download TikTok Video
      description: Download TikTok videos without watermark, including HD version and audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  description: TikTok video URL
                  example: "https://www.tiktok.com/@user/video/7123456789012345678"
      responses:
        '200':
          description: Video downloaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  video:
                    type: object
                    properties:
                      id:
                        type: string
                      title:
                        type: string
                      cover:
                        type: string
                        format: uri
                      duration:
                        type: integer
                      download:
                        type: object
                        properties:
                          no_watermark:
                            type: string
                            format: uri
                          watermark:
                            type: string
                            format: uri
                          hd:
                            type: string
                            format: uri
                      audio:
                        type: string
                        format: uri
                      stats:
                        type: object
                        properties:
                          plays:
                            type: integer
                          likes:
                            type: integer
                          comments:
                            type: integer
                          shares:
                            type: integer
                      author:
                        type: object
                        properties:
                          username:
                            type: string
                          nickname:
                            type: string
                          avatar:
                            type: string
                            format: uri
                  credits_remaining:
                    type: integer
                    example: 95
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/spotify/search:
    post:
      tags:
        - Music Services
      summary: Search Spotify Tracks
      description: Search for songs on Spotify by query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Search query (song name, artist, etc)
                  example: "I'm so paid Akon"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      query:
                        type: string
                      tracks:
                        type: array
                        items:
                          type: object
                  credits_remaining:
                    type: integer
  /api/spotify/download:
    post:
      tags:
        - Music Services
      summary: Download Spotify Track
      description: Download a track from Spotify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  example: "Bad Guy Billie Eilish"
      responses:
        '200':
          description: Download link generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  download:
                    type: object
                    properties:
                      url:
                        type: string
                        format: uri
                      title:
                        type: string
                      artist:
                        type: string
                      duration:
                        type: integer
                  credits_remaining:
                    type: integer
  /api/youtube/search:
    post:
      tags:
        - Media Downloads
      summary: Search YouTube Videos
      description: Search videos on YouTube by query (free, no external API key needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  example: "Rick Astley"
                max_results:
                  type: integer
                  minimum: 1
                  maximum: 25
                  default: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      query:
                        type: string
                      count:
                        type: integer
                      videos:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            title:
                              type: string
                            url:
                              type: string
                              format: uri
                            thumbnail:
                              type: string
                              format: uri
                            duration:
                              type: string
                            views:
                              type: string
                            published:
                              type: string
                            channel:
                              type: string
                  credits_remaining:
                    type: integer
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/youtube/download:
    post:
      tags:
        - Media Downloads
      summary: Download YouTube Video/Audio
      description: Download videos or audio from YouTube
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  example: "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                format:
                  type: string
                  enum: [video, audio]
                  default: video
                quality:
                  type: string
                  enum: [highest, lowest, 720p, 1080p]
                  default: highest
      responses:
        '200':
          description: Download ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  download:
                    type: object
                  credits_remaining:
                    type: integer
...
  /api/payment/mercadopago:
    post:
      tags:
        - Payments
      summary: Create MercadoPago Payment
      description: Create a payment link via MercadoPago (Brazil/LATAM)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - amount
                - usuario_id
              properties:
                email:
                  type: string
                  format: email
                  example: "cliente@email.com"
                amount:
                  type: number
                  minimum: 5
                  example: 50.00
                description:
                  type: string
                  example: "Compra de produto"
                usuario_id:
                  type: string
                  example: "123"
      responses:
        '200':
          description: Payment link created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
  /api/payment/mpesa:
    post:
      tags:
        - Payments
      summary: Process M-Pesa Payment
      description: Process payment via M-Pesa (Vodacom Mozambique)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - valor
                - numero_celular
                - usuario_id
              properties:
                valor:
                  type: string
                  example: "100.00"
                numero_celular:
                  type: string
                  pattern: '^(84|85)\d{7}$'
                  example: "841234567"
                usuario_id:
                  type: string
                  example: "123"
      responses:
        '200':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobilePaymentResponse'
  /api/payment/emola:
    post:
      tags:
        - Payments
      summary: Process E-Mola Payment
      description: Process payment via E-Mola (Movitel Mozambique)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - valor
                - numero_celular
                - usuario_id
              properties:
                valor:
                  type: string
                  example: "50.00"
                numero_celular:
                  type: string
                  pattern: '^(86|87)\d{7}$'
                  example: "861234567"
                usuario_id:
                  type: string
                  example: "456"
      responses:
        '200':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobilePaymentResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Get yours at: https://alauda-api.topazioverse.com.br/api/auth/register
  schemas:
...
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        provider:
          type: string
        payment:
          type: object
          properties:
            id:
              type: string
            init_point:
              type: string
              format: uri
            external_reference:
              type: string
        credits_remaining:
          type: integer
    MobilePaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        provider:
          type: string
        payment:
          type: object
          properties:
            transaction_id:
              type: string
            conversation_id:
              type: string
            response_code:
              type: string
            response_desc:
              type: string
        credits_remaining:
          type: integer
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation error"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "API Key inválida ou expirada"
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    InsufficientCredits:
      description: Not enough credits
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Créditos insuficientes"
            required: 5
            available: 2
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: "Erro interno do servidor"